+++ sample/pub_graph.py
@@ -1,6 +1,6 @@
 import numpy as np
 import math, os, random, sys, threading, time, rospy
-from std_msgs.msg import Float32
+from std_msgs.msg import Float64
 
 def plot_XY(pub_x_name, pub_y_name, pub_z_name):
 
@@ -16,8 +16,8 @@ def plot_XY(pub_x_name, pub_y_name, pub_z_name):
 			return data_y
 
 
-	pub_x = rospy.Publisher(pub_x_name, Float32, queue_size=1)
-	pub_y = rospy.Publisher(pub_y_name, Float32, queue_size=1)
+	pub_x = rospy.Publisher(pub_x_name, Float64, queue_size=1)
+	pub_y = rospy.Publisher(pub_y_name, Float64, queue_size=1)
 	# pub_z = rospy.Publisher(pub_z_name, Float32, queue_size=1)
 	rate = rospy.Rate(200) # 10hz
 	while not rospy.is_shutdown():
